@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "description": "JavaScript development environment Pluralsight course by Cory House",
   "scripts": {
-    "prestart": "node buildScripts/startMessage.js",
+    "prestart": "babel-node buildScripts/startMessage.js",
     "start": "npm-run-all --parallel security-check open:src",
-    "open:src": "node buildScripts/srcServer.js",
+    "open:src": "babel-node buildScripts/srcServer.js",
     "security-check": "nsp check",
     "localtunnel": "lt --port 3000",
     "share": "npm-run-all --parallel open:src localtunnel"
@@ -13,41 +13,41 @@
   "author": "Cory House",
   "license": "MIT",
   "dependencies": {
-    "whatwg-fetch": "1.0.0"
+    "whatwg-fetch": "2.0.2"
   },
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
-    "babel-loader": "6.2.5",
-    "babel-preset-latest": "6.16.0",
-    "babel-register": "6.16.3",
+    "babel-cli": "6.23.0",
+    "babel-core": "6.23.1",
+    "babel-loader": "6.3.2",
+    "babel-preset-latest": "6.22.0",
+    "babel-register": "6.23.0",
     "chai": "3.5.0",
     "chalk": "1.1.3",
     "cheerio": "0.22.0",
     "compression": "1.6.2",
-    "cross-env": "3.1.3",
-    "css-loader": "0.25.0",
-    "eslint": "3.8.1",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-watch": "2.1.14",
-    "express": "4.14.0",
-    "extract-text-webpack-plugin": "1.0.1",
-    "html-webpack-plugin": "2.22.0",
-    "jsdom": "9.8.0",
-    "json-schema-faker": "0.3.6",
-    "json-server": "0.8.22",
-    "localtunnel": "1.8.1",
-    "mocha": "3.1.2",
-    "nock": "8.1.0",
-    "npm-run-all": "3.1.1",
+    "cross-env": "3.1.4",
+    "css-loader": "0.26.1",
+    "eslint": "3.15.0",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-watch": "3.0.0",
+    "express": "4.14.1",
+    "extract-text-webpack-plugin": "2.0.0-beta.4",
+    "html-webpack-plugin": "2.28.0",
+    "jsdom": "9.11.0",
+    "json-schema-faker": "0.4.0",
+    "json-server": "0.9.5",
+    "localtunnel": "1.8.2",
+    "mocha": "3.2.0",
+    "nock": "9.0.6",
+    "npm-run-all": "4.0.1",
     "nsp": "2.6.2",
-    "numeral": "1.5.3",
+    "numeral": "2.0.4",
     "open": "0.0.5",
-    "rimraf": "2.5.4",
+    "rimraf": "2.6.1",
     "style-loader": "0.13.1",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.8.4",
-    "webpack-hot-middleware": "2.13.0",
+    "webpack": "2.2.1",
+    "webpack-dev-middleware": "1.10.1",
+    "webpack-hot-middleware": "2.17.0",
     "webpack-md5-hash": "0.0.5"
   }
 }
