@@ -1,3 +1,4 @@
-var chalk = require('chalk');
+//var chalk = require('chalk');
+import chalk from 'chalk';
 
 console.log(chalk.green('Start app in dev mode...'));
